% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clickstreams}
\alias{clickstreams}
\title{Sample Clickstream Data}
\format{
Data prepared for using the \code{"clickstream"} R package.
}
\usage{
data(clickstreams)
}
\description{
Devised clickstream data to display the visualisations
and show the interactions between the clickstream package.
}
\examples{
mc <- fitMarkovChain(clickstreams)
plot_transition_matrix(mc)
}
\keyword{dataset}
