% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_data.R
\docType{data}
\name{clickstreams_large}
\alias{clickstreams_large}
\title{Larger sample of clickstream data}
\format{
Data prepared by using the \code{"clickstream"} R package.
}
\usage{
data(clickstreams_large)
}
\description{
Devised clickstream data for examples of the visualisations.
}
\examples{
plot_transition_matrix(clickstreams_large)
}
\keyword{dataset}
