% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transition_graph.R
\name{create_transition_graph}
\alias{create_transition_graph}
\title{Create Markov chain transition network}
\usage{
create_transition_graph(data, state_labels)
}
\arguments{
\item{data}{Fitted Markov chain}

\item{state_labels}{Threshold for displaying transition probabilities.}
}
\value{
Network visualisation
}
\description{
Builds a network graph to represent the transition matrix.
}
\examples{
data <- readClickstreams(file)
create_transition_graph(data)
}
