% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transition_network.R
\name{create_transition_network}
\alias{create_transition_network}
\title{Create the transition network}
\usage{
create_transition_network(data, vertex_labels)
}
\arguments{
\item{data}{Clickstream data}

\item{vertex_labels}{A dataframe of labelled states (state, vertex_label)}
}
\value{
Network visualisation of the transition matrix
}
\description{
A quick plot of the first order Markov chain transition matrix.
}
\examples{
cls <- readClickstreams(file)
create_transition_network(cls, vertex_labels)
}
