% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot Markov chain transition network}
\usage{
plot_graph(
  graph_edges,
  graph_nodes,
  node_size_factor = 1.2,
  edge_width_factor = 5,
  edge_arrow_size = 0.6,
  edge_curve_factor = 0.5,
  legend_x_position = -3,
  legend_y_position = -0.5
)
}
\arguments{
\item{graph_edges}{Fitted Markov chain}

\item{graph_nodes}{Threshold for displaying transition probabilities.}

\item{node_size_factor}{Threshold for displaying transition probabilities.}

\item{edge_width_factor}{Threshold for displaying transition probabilities.}

\item{edge_arrow_size}{Threshold for displaying transition probabilities.}

\item{edge_curve_factor}{Threshold for displaying transition probabilities.}

\item{legend_x_position}{Threshold for displaying transition probabilities.}

\item{legend_y_position}{Threshold for displaying transition probabilities.}
}
\value{
Network visualisation
}
\description{
Builds a network graph to represent the transition matrix.
}
\examples{
cls <- readClickstreams(file)
mc <- fitMarkovChain(cls)
plot_graph(mc)
}
