% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transition_matrix.R
\name{plot_transition_heatmap}
\alias{plot_transition_heatmap}
\title{Plot transition heatmap}
\usage{
plot_transition_heatmap(data)
}
\arguments{
\item{data}{Clickstream data}
}
\value{
Heatmap of the transition matrix
}
\description{
Builds an interactive heatmap of the first order Markov chain transition matrix.
}
\examples{
cls <- readClickstreams(file)
plot_transition_heatmap(cls)
}
