% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transition_matrix.R
\name{plot_transition_matrix}
\alias{plot_transition_matrix}
\title{Plot transition matrix}
\usage{
plot_transition_matrix(data)
}
\arguments{
\item{data}{Clickstream data}
}
\value{
Heatmap of the transition matrix
}
\description{
Builds an interactive heatmap of the first order Markov chain transition matrix.
}
\examples{
cls <- readClickstreams(file)
plot_transition_matrix(cls)
}
