% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transition_network.R
\name{plot_transition_network}
\alias{plot_transition_network}
\title{Plot Markov chain transition network}
\usage{
plot_transition_network(
  mc,
  lower_bound = 0.01,
  state_hits = c(127, 830, 557, 9800, 4749, 862, 620, 65, 6323, 998, 66, 572, 634, 235),
  labels = factor(c("Content", "Purchasing", "Content", "Products", "Artificial",
    "Home", "Site Information", "Profile", "Products", "Artificial", "Products",
    "Profile", "Purchasing", "Purchasing")),
  node_size_factor = 0.06,
  edge_width_factor = 10,
  edge_arrow_size = 0.8,
  edge_curved = 0.6,
  layout_size = 1.5,
  legend_x = -2,
  legend_y = -0.5
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Plot Markov chain transition network
}
\examples{
cls <- readClickstreams(file)
mc <- fitMarkovChain(cls)
plot_transition_network(mc)


}
