% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_graph_edges.R
\name{prepare_graph_edges}
\alias{prepare_graph_edges}
\title{Generate network graph edge data}
\usage{
prepare_graph_edges(data, lower_bound = 0.01)
}
\arguments{
\item{lower_bound}{Threshold for displaying transition probabilities.}

\item{mc}{Fitted Markov chain}
}
\value{
Network visualisation
}
\description{
Builds a network graph to represent the transition matrix.
}
\examples{
cls <- readClickstreams(file)
mc <- fitMarkovChain(cls)
prepare_graph_edges(mc)
}
