% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_graph_nodes.R
\name{prepare_graph_nodes}
\alias{prepare_graph_nodes}
\title{Generate network graph node data}
\usage{
prepare_graph_nodes(data, state_labels = state_labels_ex)
}
\arguments{
\item{data}{Fitted Markov chain}

\item{state_labels}{Threshold for displaying transition probabilities.}
}
\value{
Network visualisation
}
\description{
Builds a network graph to represent the transition matrix.
}
\examples{
cls <- readClickstreams(file)
mc <- fitMarkovChain(cls)
prepare_graph_nodes(mc)
}
