% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_network_edges.R
\name{prepare_network_edges}
\alias{prepare_network_edges}
\title{Prepare the network edges}
\usage{
prepare_network_edges(data, alpha = 0.01)
}
\arguments{
\item{data}{Clickstream data}

\item{alpha}{Display only p_{i,j} > alpha}
}
\value{
Network edges data
}
\description{
Transform and clean the clickstream data into transition probabilites, p_{i,j}.
Output dataframe of network edges.
}
\examples{
cls <- readClickstreams(file)
prepare_network_edges(cls)
}
