% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_network_vertices.R
\name{prepare_network_vertices}
\alias{prepare_network_vertices}
\title{Prepare the network vertices}
\usage{
prepare_network_vertices(data, vertex_labels)
}
\arguments{
\item{data}{Clickstream data}

\item{vertex_labels}{A dataframe of labelled states (state, vertex_label)}
}
\value{
Network vertices data
}
\description{
Find the frequency of each state in the clickstream data.
Output dataframe of network vertices and labels.
}
\examples{
cls <- readClickstreams(file)
prepare_network_vertices(cls, vertex_labels)
}
