% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clickstreams}
\alias{clickstreams}
\title{Sample clickstream data}
\format{
Data prepared for using the \code{"clickstream"} R package.
}
\usage{
data(clickstreams)
}
\description{
Devised clickstream data for examples of the visualisations.
}
\examples{
plot_transition_matrix(data)
}
\keyword{dataset}
