% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clickstreams}
\alias{clickstreams}
\title{Ryan's Sample Data}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(clickstreams)
}
\description{
Data made up of clickstream with various states.
}
\examples{
mc <- fitMarkovChain(clickstreams)
plot_transition_matrix(mc)
}
\keyword{datasets}
