% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transition_matrix.R
\name{plot_transition_matrix}
\alias{plot_transition_matrix}
\title{Plot Markov chain transition matrix}
\usage{
plot_transition_matrix(mc)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Builds a heatmap of the Markov chain transition matrix.
}
\examples{
cls <- readClickstreams(file)
mc <- fitMarkovChain(cls)
plot_transition_matrix(mc)
}
