% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transition_network.R
\name{plot_transition_network}
\alias{plot_transition_network}
\title{Plot the transition network}
\usage{
plot_transition_network(
  network_edges,
  network_vertices,
  edge_width_factor = 5,
  edge_arrow_size = 0.6,
  edge_curve_factor = 0.5,
  vertex_size_factor = 1.2,
  legend_x_position = -3,
  legend_y_position = -0.5
)
}
\arguments{
\item{network_edges}{Network edges data}

\item{network_vertices}{Network vertices data}

\item{edge_width_factor}{Transform the edge width values}

\item{edge_arrow_size}{Set the arrow size}

\item{edge_curve_factor}{Set the edge curve}

\item{vertex_size_factor}{Transform the vertex size values}

\item{legend_x_position}{x-axis value for the legend}

\item{legend_y_position}{y-axis value for the legend}
}
\value{
Network visualisation of the transition matrix
}
\description{
Generate a network to represent the first order Markov chain transition matrix.
}
\examples{
cls <- readClickstreams(file)
edges <- prepare_network_edges(cls)
vertices <- prepare_network_vertices(cls, vertex_labels)
plot_transition_network(edges, vertices)
}
